+++ opensearch_py_ml/ml_commons/ml_commons_client.py
@@ -13,7 +13,7 @@ from typing import Any, List, Optional, Union
 from deprecated.sphinx import deprecated
 from opensearchpy import OpenSearch
 
-from opensearch_py_ml.ml_commons.connector import Connector
+from opensearch_py_ml.ml_commons.model_connector import Connector
 from opensearch_py_ml.ml_commons.ml_common_utils import (
     ML_BASE_URI,
     MODEL_FORMAT_FIELD,
